@@ -1,47 +1,47 @@
 {
   "name": "dashboard",
-  "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "web-vitals": "^1.1.0"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.7.1",
-    "@vitejs/plugin-react": "^4.3.4",
-    "vite": "^6.0.3",
-    "jest": "^29.0.0"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "lint": "eslint .",
     "preview": "vite preview",
     "test": "jest"
   },
-  "eslintConfig": {
-    "extends": ["react-app", "react-app/jest"]
+  "dependencies": {
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "@babel/preset-env": "^7.18.10",
+    "@babel/preset-react": "^7.18.6",
+    "@eslint/js": "^9.15.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "eslint": "^9.15.0",
+    "eslint-plugin-react": "^7.37.2",
+    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-refresh": "^0.4.14",
+    "globals": "^15.12.0",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.2.0",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/user-event": "^14.2.1",
+    "jest-environment-jsdom": "^29.2.0",
+    "vite": "^6.0.1"
   },
   "jest": {
-    "testEnvironment": "jsdom",
-    "moduleNameMapper": {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/fileTransformer.js",
-      "^.+\\.(css|less|scss)$": "identity-obj-proxy"
-    },
-    "setupFilesAfterEnv": ["./setupTests.js"]
-  }
+  "testEnvironment": "jsdom",
+  "moduleNameMapper": {
+    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/fileTransformer.js",
+    "^.+\\.(css|less|scss)$": "identity-obj-proxy"
+  },
+  "setupFilesAfterEnv": [
+    "./setupTests.js"
+  ]
+}
 }
