@@ -1,22 +1,15 @@
-const getFullYear = () => {
-	let current_year = new Date().getFullYear();
-	return current_year;
-}
+// task_1/dashboard/src/utils.js
 
-const getFooterCopy = (isIndex) => {
-	if (isIndex === true) {
-		return "Holberton School";
-	} else {
-		return "Holberton School main dashboard";
-	}
-}
-
-const getLatestNotification = () => {
-	return "<strong>Urgent requirement</strong> - complete by EOD";
-};
-
-module.exports = {
-	getFooterCopy,
-	getFullYear,
-	getLatestNotification,
-};
+// Function to get the current year
+export function getCurrentYear() {
+    return new Date().getFullYear();
+  }
+  
+  // Function to get footer copy based on whether it's the index page or not
+  export function getFooterCopy(isIndex) {
+    if (isIndex) {
+      return 'Holberton School';
+    } else {
+      return 'Holberton School main dashboard';
+    }
+  }
